refactor : reRelandering Prevention
