+++ pages/detail/[id].tsx
@@ -38,6 +38,7 @@ import KebabMenuUI, {
   KebabMenuStyle,
   KebabMenuAptionButton,
 } from '@/components/common/KebabMenuUI';
+import DetailKebabMenu from '@/components/Detail/DetailKebabMenu';
 import { getComment, reportPost } from '@/api/post';
 import { identity } from 'lodash';
 import Link from 'next/link';
@@ -75,65 +76,7 @@ const Detail: NextPage = () => {
   const [videoId, setVideoId] = useState<string>('');
   const [tags, setTags] = useState<string[]>([]);
   const [location, setLocation] = useState({ name: '', x: 0, y: 0 });
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-  const [modalMessge, setModalMessge] = useState<any>({
-    actionText: '',
-    modalMessge: '',
-    onClickEvent: '',
-  });
-  const {
-    isModalOpen: oneModalOpen,
-    modalMessge: oneModalMessage,
-    openModal: openOneModal,
-    closeModal: closeOneModal,
-  } = useModal();
-  const queryClient = useQueryClient();
 
-  const { mutate: delete_mutate } = useMutation(deletePost, {
-    onSuccess: () => {
-      queryClient.invalidateQueries('posts');
-
-      openOneModal({
-        actionText: '확인',
-        modalMessge: '게시물이 삭제 되었습니다.',
-        onClickEvent: () => router.push('/'), // 이 부분을 추가합니다.
-      });
-    },
-  });
-
-  const { mutate: report } = useMutation(reportPost, {
-    onSuccess: (messag) => {
-      setModalMessge({
-        actionText: '확인',
-        modalMessge: messag,
-        onClickEvent: null,
-      });
-      setIsModalOpen(true);
-    },
-  });
-
-  const handleEditClick = () => {
-    router.push(`/update/${id}`);
-  };
-
-  const handleDeleteClick = () => {
-    setModalMessge({
-      actionText: '삭제',
-      modalMessge: '정말로 해당 게시글을 삭제하시겠습니까?',
-      onClickEvent: () => delete_mutate(id),
-    });
-    setIsModalOpen(true);
-  };
-
-  const handleReportClick = () => {
-    setModalMessge({
-      actionText: '신고',
-      modalMessge: '해당 사용자를 신고하시겠습니까?',
-      onClickEvent: () => report(id),
-    });
-    setIsModalOpen(true);
-  };
-  // console.log(isModalOpen);
   const isClientSide = typeof window !== 'undefined';
   const tokenExists = isClientSide ? !!localStorage.getItem('token') : false;
 
@@ -183,24 +126,6 @@ const Detail: NextPage = () => {
   if (!isSuccess) return <div>Loading...</div>;
   return (
     <>
-      {/* <GlobalStyle /> */}
-      {oneModalOpen && (
-        <OneButtonModal
-          isOpen={oneModalOpen}
-          setIsOpen={closeOneModal}
-          onClickEvent={oneModalMessage.onClickEvent || null}
-        >
-          {oneModalMessage.modalMessge}
-        </OneButtonModal>
-      )}
-      <Modal
-        isOpen={isModalOpen}
-        setIsOpen={setIsModalOpen}
-        actionText={modalMessge.actionText}
-        onClickEvent={modalMessge.onClickEvent}
-      >
-        {modalMessge.modalMessge}
-      </Modal>
       <ContentsContainer>
         <OtherProfileBox>
           <Link href={`/User/${data?.user?.userId}`}>
@@ -210,24 +135,11 @@ const Detail: NextPage = () => {
               nick={data?.user?.nickname}
             />
           </Link>
-          <KebabMenuUI>
-            <KebabMenuStyle>
-              {data.user.email === userData?.email ? (
-                <>
-                  <KebabMenuAptionButton onClick={handleDeleteClick}>
-                    글 삭제하기 <span></span>
-                  </KebabMenuAptionButton>
-                  <KebabMenuAptionButton onClick={handleEditClick}>
-                    글 수정하기 <span></span>
-                  </KebabMenuAptionButton>
-                </>
-              ) : (
-                <KebabMenuAptionButton onClick={handleReportClick}>
-                  신고하기 <span></span>
-                </KebabMenuAptionButton>
-              )}
-            </KebabMenuStyle>
-          </KebabMenuUI>
+          <DetailKebabMenu
+            data={data?.user.email}
+            userData={userData?.email}
+            id={id}
+          />
         </OtherProfileBox>
         <div className="carousel-box">
           <MainBannerSlider data={images} />
